+++ eslint.config.js
@@ -24,7 +24,7 @@ export default tseslint.config({ ignores: ['dist'] }, eslintConfigPrettier, {
       'warn',
       { allowConstantExport: true },
     ],
-    'no-undef': 'warn',
+    'no-undef': 'error',
     'no-console': 'error',
     'react/no-array-index-key': 'error',
   },
