+++ src/types/classNames.ts
@@ -1,0 +1,53 @@
+export type ClassNames = {
+  dateRangeWrapper?: string | undefined;
+  calendarWrapper?: string | undefined;
+  dateDisplay?: string | undefined;
+  dateDisplayItem?: string | undefined;
+  dateDisplayItemActive?: string | undefined;
+  monthAndYearWrapper?: string | undefined;
+  monthAndYearPickers?: string | undefined;
+  nextPrevButton?: string | undefined;
+  month?: string | undefined;
+  weekDays?: string | undefined;
+  weekDay?: string | undefined;
+  days?: string | undefined;
+  day?: string | undefined;
+  dayNumber?: string | undefined;
+  dayPassive?: string | undefined;
+  dayToday?: string | undefined;
+  dayStartOfWeek?: string | undefined;
+  dayEndOfWeek?: string | undefined;
+  daySelected?: string | undefined;
+  dayDisabled?: string | undefined;
+  dayStartOfMonth?: string | undefined;
+  dayEndOfMonth?: string | undefined;
+  dayWeekend?: string | undefined;
+  dayStartPreview?: string | undefined;
+  dayInPreview?: string | undefined;
+  dayEndPreview?: string | undefined;
+  dayHovered?: string | undefined;
+  dayActive?: string | undefined;
+  inRange?: string | undefined;
+  endEdge?: string | undefined;
+  startEdge?: string | undefined;
+  prevButton?: string | undefined;
+  nextButton?: string | undefined;
+  selected?: string | undefined;
+  months?: string | undefined;
+  monthPicker?: string | undefined;
+  yearPicker?: string | undefined;
+  dateDisplayWrapper?: string | undefined;
+  definedRangesWrapper?: string | undefined;
+  staticRanges?: string | undefined;
+  staticRange?: string | undefined;
+  inputRanges?: string | undefined;
+  inputRange?: string | undefined;
+  inputRangeInput?: string | undefined;
+  dateRangePickerWrapper?: string | undefined;
+  staticRangeLabel?: string | undefined;
+  staticRangeSelected?: string | undefined;
+  monthName?: string | undefined;
+  infiniteMonths?: string | undefined;
+  monthsVertical?: string | undefined;
+  monthsHorizontal?: string | undefined;
+};
